backend/requirements.txt
fastapi
uvicorn
python-multipart
fastapi-mail
pydantic
groq
python-dotenv
requests
motor
certifi
python-jose
